fix : settings, signout
