+++ src/main/java/com/example/itDa/domain/model/Community.java
@@ -33,14 +33,8 @@ public class Community {
 
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = true)
+    @JoinColumn(name = "user_id")
     private User user;
-
-    @OneToMany(mappedBy = "community", fetch = FetchType.LAZY, cascade = CascadeType.ALL)
-    private List<Comment> comments = new ArrayList<>();
-
-    @OneToMany(mappedBy = "community", fetch = FetchType.LAZY, cascade = CascadeType.ALL)
-    private List<CommunityFile> communityFiles;
 
     public void updateCommunity(CommunityRequestDto requestDto){
         this.title = requestDto.getTitle();
