+++ src/main/java/com/example/itDa/domain/repository/ArticleRepository.java
@@ -1,16 +1,10 @@
-package com.example.itDa.domain.article.repository;
+package com.example.itDa.domain.repository;
 
-import com.example.itDa.domain.Category;
-import com.example.itDa.domain.article.Article;
+import com.example.itDa.domain.model.Article;
 import com.example.itDa.domain.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 
 import java.util.List;
-import java.util.Optional;
-
-
 
 
 public interface ArticleRepository extends JpaRepository<Article, Long>, ArticleRepositoryCustom {
