+++ src/main/java/com/example/itDa/domain/repository/ArticleRepositoryCustom.java
@@ -1,11 +1,9 @@
-package com.example.itDa.domain.article.repository;
+package com.example.itDa.domain.repository;
 
-import com.example.itDa.domain.article.response.ViewAllArticleResponseDto;
+import com.example.itDa.dto.response.ViewAllArticleResponseDto;
 import com.example.itDa.infra.security.UserDetailsImpl;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-
-import java.util.List;
 
 public interface ArticleRepositoryCustom {
     Page<ViewAllArticleResponseDto> getViewAllArticle(Pageable pageable, UserDetailsImpl userDetails);
