+++ src/main/java/com/example/itDa/dto/request/EditArticleRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.itDa.domain.article.request;
+package com.example.itDa.dto.request;
 
 import com.example.itDa.domain.Category;
 import lombok.Builder;
