+++ src/main/java/com/example/itDa/dto/response/ArticleResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.itDa.domain.article.response;
+package com.example.itDa.dto.response;
 
 import com.example.itDa.domain.Category;
 import com.example.itDa.domain.Status;
@@ -12,17 +12,24 @@ import java.util.List;
 
 @Getter
 @Builder
-@AllArgsConstructor
 @NoArgsConstructor
-public class EditArticleResponseDto {
+@AllArgsConstructor
+public class ArticleResponseDto {
+    private Long userId;
+    private String username;
+    private int density;
     private Long articleId;
     private String articleName;
     private String substance;
     private Category category;
     private Status status;
     private String location;
+    private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
     private int sellPrice;
     private List<String> fileName;
     private List<String> fileUrl;
+    private boolean like;
+
+
 }
